add `hovered = false` to `TableRow.getDefaultProps()` (+1 squashed commits)
Squashed commits:

[30bfb46] [TableRow] add support for `this.props.hovered` in addition to `this.state.hovered`

Allow specifying the `hovered` state for a row via `props` in addition to the internal `state`.

This can be used to synchronize the hovered information with some other external events, e.g., for linking a grid view with a table view highlighting the same items on hovering in either view.