[core] Batch small changes (#18155)
* [docs] Link https://material-ui.com/versions/

This will enable developer landing on an old version to easy find its way to
a different version.

The idea was sparked by https://www.ag-grid.com/archive/.

* [examples] Remove parcel

I have added a Parcel demo 18 months ago, anticipating it would gain traction.
https://github.com/mui-org/material-ui/pull/10575

As it turns out, I was wrong. I think that it's too early to have an official example with Parcel.
Our example has outdated dependencies.

https://npm-stat.com/charts.html?package=parcel-bundler&package=preact&package=webpack&from=2017-11-01&to=2019-11-01

For instance, Preact has more traction, it would be better to improve our example for it.

* [docs] Add missing 180px icon

* Thank you Tom

* [docs] Remove Blokt backer

* [docs] Remove Book affiliation

This is not interesting. It yields about $5/month.

* [TableCell] Smaller source

* [CHANGELOG] Fix typos

* [docs] Use rel=sponsored where appropriate

Bring to light by Sebastien, for more details:
https://moz.com/blog/nofollow-sponsored-ugc

* [docs] Reduce diffs, isolate changes

* [core] Prefer parentNode over parentElement

They are very close, it seems that the difference will never impact us.
https://stackoverflow.com/questions/8685739/difference-between-dom-parentnode-and-parentelement

I went with parentNode as more frequent in the codebase.

* [docs] Stop date picker survey

* [Autocomplete] Add active custom state